plugin-drizzle: fix state persistence when switching back to backfill
